authetication/requirements.txt
asgiref==3.4.1; python_version >= "3.8"
django==4.0; python_version >= "3.8"
djangorestframework-simplejwt==5.0.0; python_version >= "3.7"
djangorestframework==3.13.0; python_version >= "3.7"
gunicorn==20.1.0; python_version >= "3.5"
pyjwt==2.3.0; python_version >= "3.7"
pytz==2021.3; python_version >= "3.7"
sqlparse==0.4.2; python_version >= "3.8"
tzdata==2021.5; sys_platform == "win32" and python_version >= "3.8"
whitenoise==5.3.0; python_version >= "3.5" and python_version < "4"
